requirements.txt
boto3
cryptography
tksheet
pandas
Pillow